+++ Add Code Here/PYTHON/Factorial.py
@@ -1,7 +1,19 @@
-def factorial(n):
-    return 1 if (n==1 or n==0) else n * factorial(n - 1);
+# Python program to find the factorial of a number provided by the user.
 
-number = int(input("Enter a number: "))
-print("Factorial of " + str(number) + " is",
-factorial(number))
-# Ordinary-Droid
+# change the value for a different result
+num = 7
+
+# To take input from the user
+#num = int(input("Enter a number: "))
+
+factorial = 1
+
+# check if the number is negative, positive or zero
+if num < 0:
+   print("Sorry, factorial does not exist for negative numbers")
+elif num == 0:
+   print("The factorial of 0 is 1")
+else:
+   for i in range(1,num + 1):
+       factorial = factorial*i
+   print("The factorial of",num,"is",factorial)
